+++ src/jsonreflow/reflow.py
@@ -1,8 +1,21 @@
+import contextlib
 import json
-from typing import Iterable, Iterator, List
+import os
+import tempfile
+from pathlib import Path
+from typing import Iterable, Iterator, List, Protocol, Union
 
 MAX_WIDTH_DEFAULT = 80
 INDENT_DEFAULT = 2
+
+
+class SupportsRead(Protocol):
+    def read(self, size: int = ..., /) -> str: ...
+    def readlines(self) -> Iterable[str]: ...
+
+
+class SupportsWrite(Protocol):
+    def write(self, text: str, /) -> int: ...
 
 
 def reflow_iter(
@@ -61,6 +74,9 @@ def reflow_iter(
 
 
 def reflow(encoded: str, *, max_width: int = MAX_WIDTH_DEFAULT) -> str:
+    """
+    Reflow the given encoded JSON string.
+    """
     return "\n".join(
         reflow_iter(
             lines=encoded.split("\n"),
@@ -104,6 +120,17 @@ def _chunks_to_lines(chunks: Iterable[str]) -> Iterator[str]:
         yield buffer
 
 
+def _json_encode_lines(obj, indent: int = INDENT_DEFAULT) -> Iterator[str]:
+    """
+    Use stdlib json.JSONEncoder to JSON-encode given object
+    and produce line per line
+    """
+    # TODO: support all/most of JSONEncoder's arguments?
+    encoder = json.JSONEncoder(indent=indent)
+    chunks = encoder.iterencode(obj)
+    yield from _chunks_to_lines(chunks)
+
+
 def dump(
     obj,
     fp,
@@ -118,9 +145,80 @@ def dump(
     but with reflowing to fit within a given line width.
     """
     # TODO: support all/most of JSONEncoder's arguments?
-    encoder = json.JSONEncoder(indent=indent)
-    chunks = encoder.iterencode(obj)
-    lines = reflow_iter(lines=_chunks_to_lines(chunks), max_width=max_width)
-    for line in lines:
+    encoded_lines = _json_encode_lines(obj=obj, indent=indent)
+    for line in reflow_iter(lines=encoded_lines, max_width=max_width):
         # TODO: classic json.dump() does not add newline after last line
         fp.write(line + "\n")
+
+
+@contextlib.contextmanager
+def _temp_sink_and_rename_on_exit(
+    path: Path, *, mode: str = "w", encoding: str = "utf-8"
+):
+    """
+    Context manager for safe in-place writing of a file
+    (avoid truncating the original file too early):
+    use a temporary file during writing,
+    and atomically replace the target path on successful exit of the context manager.
+    """
+
+    # Use temp file in same directory (file system) to allow atomic move
+    folder = path.parent
+
+    with tempfile.NamedTemporaryFile(
+        mode=mode,
+        encoding=encoding,
+        prefix=".jsonreflow_tmp_",
+        dir=folder,
+        delete=False,
+    ) as temp_file:
+        try:
+            yield temp_file
+            temp_file.flush()
+            temp_file.close()
+            # Rename to target path (should be atomic move)
+            os.replace(src=temp_file.name, dst=path)
+        except Exception:
+            # Clean up temp file on error
+            with contextlib.suppress(FileNotFoundError):
+                os.remove(temp_file.name)
+            raise
+
+
+def reflow_file(
+    input: Union[str, Path, SupportsRead],
+    output: Union[str, Path, SupportsWrite, None],
+    *,
+    assume_formatted: bool = False,
+    max_width: int = MAX_WIDTH_DEFAULT,
+    indent: int = INDENT_DEFAULT,
+) -> None:
+    """
+    Reflow JSON from the given file and write back in-place
+    or write to another file.
+    """
+    if isinstance(input, (str, Path)):
+        input_context = Path(input).open(mode="r", encoding="utf-8")  # noqa: SIM115
+    else:
+        # Assume it's already a readable file-like object
+        input_context = contextlib.nullcontext(input)
+
+    if output is None:
+        # In-place mode
+        assert isinstance(input, (str, Path))
+        output_context = _temp_sink_and_rename_on_exit(path=Path(input))
+    elif isinstance(output, (str, Path)):
+        output_context = _temp_sink_and_rename_on_exit(path=Path(output))
+    else:
+        # Assume it's already a writable file-like object
+        output_context = contextlib.nullcontext(output)
+
+    with output_context as output_file, input_context as input_file:
+        if assume_formatted:
+            encoded_lines = (s.rstrip() for s in input_file.readlines())
+        else:
+            data = json.load(fp=input_file)
+            encoded_lines = _json_encode_lines(obj=data, indent=indent)
+
+        for line in reflow_iter(lines=encoded_lines, max_width=max_width):
+            output_file.write(line + "\n")